added all important repositories

+++ app/backend/src/main/java/ap/student/project/backend/dao/CertificateRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CertificateRepository extends ListCrudRepository<Certificate, Long> {
+public interface CertificateRepository extends ListCrudRepository<Certificate, Integer> {
 }
