+++ app/backend/src/main/java/ap/student/project/backend/dao/ChatRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChatRepository extends ListCrudRepository<Chat, Long> {
+public interface ChatRepository extends ListCrudRepository<Chat, Integer> {
 }
